Split cpuid.go into generic and amd64-specific code
